+++ workspace/src/solvers/jacobipthreadsolver.h
@@ -1,7 +1,3 @@
-//
-// Created by caos on 06/06/17.
-//
-
 #ifndef JACOBI_JACOBIPTHREADSOLVER_H
 #define JACOBI_JACOBIPTHREADSOLVER_H
 
@@ -12,14 +8,14 @@
 class JacobiPThreadSolver : public JacobiSolver
 {
 public:
-    JacobiPThreadSolver(const double** A, const double* b, int N, int nWorkers);
+    JacobiPThreadSolver(const float** A, const float* b, int N, int nWorkers);
 
 private:
     int mnWorkers;
 
-    void deltax(const double *x, double *dest);
+    void deltax(const float *x, float *dest);
     int getNWorkers() { return mnWorkers; };
-    void threadBody(int minI, int maxI, const double* x, double* dest);
+    void threadBody(int minI, int maxI, const float* x, float* dest);
 
 };
